# s.c.i.Hash{Map,Set} concat/removedAll use shallow mutations