… src/library/scala/collection/immutable/HashMap.scala
… src/library/scala/collection/immutable/HashSet.scala