## Summary

### TL;DR

* immutable HashMaps and HashSets are tree structures
* When performing `concat` on them, previously we would immutably add the argument's elements to the Map/Set, one at a time, allocating a new node for each node on the path to the changed key
* Now, we use mutation when performing further changes, once a new node has been created at the root level, or at the level immediately below the root.
* Same thing for `removedAll`, except that method only performs the trick on the root level, not the level beneath root.
* `HashMap#concat`'s performance is improved enormously, `HashSet#concat`, `HashMap#removedAll`, `HashSet#removedAll` less so, but still noticably. See benchmarks below.


### Longer explanation

When adding multiple elements to a HashMap/HashSet, we typically do something like
```scala
val iter = collection.iterator
var currentNode = rootNode
while(iter.hasNext) {
  currentNode = currentNode.updated(iter.next())
}
return new HashMap(currentNode)
```
However, this PR exploits the fact that when calculating an updated rootNode for a HashMap, if the returned node is not reference-equal to the existing node, then it must be a newly created node, and therefor has not been returned to users yet, and is therefor safe to mutate in further updates. That is, the root node is able to be mutated, but not necessarily are any of its children.

So the above becomes

```scala
      val iter = that.iterator
      var current = rootNode.asInstanceOf[BitmapIndexedMapNode[K, V1]]
      while (iter.hasNext) {
        val (key, value) = iter.next()
        val originalHash = key.##
        val improved = improve(originalHash)
        current = current.updated(key, value, originalHash, improved, 0).asInstanceOf[BitmapIndexedMapNode[K, V]]

        if (current ne rootNode) {
          while (iter.hasNext) {
            val (key, value) = iter.next()
            val originalHash = key.##
            val improved = improve(originalHash)
            // new special-purpose `update` implementation which can/will mutate only root node.
            current.updateWithShallowMutations(key, value, originalHash, improved, 0)
          }
          return new HashMap(current)
        }
      }
      this
```


We then take this process one level deeper. the calling method `HashMap#concat` can cooperate with the mutating method to keep track of which immediate node-children of the root have so far been newly created. These are tracked in a bitmap called `shallowlyMutableNodeMap: Int`. This is fed into the `updateWithShallowMutations` method. Any `1` bits in that bitmap give the method permission to mutate the node at that position. 

The `updateWithShallowMutations` will return an updated bitmap including any newly created child nodes. Then in subsequent calls to that method, those newly-created nodes will be able to be mutated. 

So the final algorithm looks like:
```scala
      val iter = that.iterator
      var current = rootNode.asInstanceOf[BitmapIndexedMapNode[K, V1]]
      while (iter.hasNext) {
        val (key, value) = iter.next()
        val originalHash = key.##
        val improved = improve(originalHash)
        current = current.updated(key, value, originalHash, improved, 0).asInstanceOf[BitmapIndexedMapNode[K, V]]

        if (current ne rootNode) {
          assert(current.content ne rootNode.asInstanceOf[BitmapIndexedMapNode[K, V]].content)
          var shallowlyMutableNodeMap = 0
          while (iter.hasNext) {
            val (key, value) = iter.next()
            val originalHash = key.##
            val improved = improve(originalHash)
            shallowlyMutableNodeMap = current.updateWithShallowMutations(key, value, originalHash, improved, 0, shallowlyMutableNodeMap)
          }
          return new HashMap(current)
        }
      }
```

Why only two levels deep? Because it's pretty easy to keep track of a single `Int` bitmap, but any further would require allocating an `Array[Int](32)` which is not as clearly a win.

## Benchmarks

(removed because they were outdated. see below)

## Other changes:

* To avoid lots of casting to `BitMapIndexedMapNode[K, V]`, `HashMap#rootNode` has been refined to `BitMapIndexedMapNode[K, V]`. It always was that type at runtime, but now it is statically known.

## TODO

- [x] Perform same optimizations for HashMap#removedAll
- [x] Perform same optimizations for HashSet#concat
- [x] Perform same optimizations for HashSet#removedAll
- [x] Special tests to ensure previously exposed nodes are not mutated due to programming errors?
- [x] Extensive scaladocs on all added/changed methods  
- [x] Benchmark operations on HashSet