@joshlemer 2019-02-05
I see, there's a bug in my code, some pattern matches are not exhaustive.