@joshlemer 2019-02-05
Marking as 2.13.0 because the performance gains are very significant