@joshlemer 2019-02-17
I think the way to go is to not try to perform the tricky shifting logic in HashMap's, because the logic is complicated, and affects other things like iteration and equality, and the benefit is slim if it even exists.

For HashSets however, there's a much stronger case to be made. Since nodes and payloads both take the same number of slots, migrating data to inline or to a node is always done without changing the size of the content array, so the logic is simple, deterministic (no if-branching), self-contained in the `migrateFromInlineToNodeInPlace` method, and does not affect anything else like equality etc.

There is a noticable improvement from at least the 50-5000 size ranges of HashSet concatenation, in CPU time and GC allocation



```scala
@BenchmarkMode(Array(Mode.AverageTime))
@Fork(2)
@Threads(1)
@Warmup(iterations = 10)
@Measurement(iterations = 10)
@OutputTimeUnit(TimeUnit.NANOSECONDS)
@State(Scope.Benchmark)
class HashSetBenchmark {
  @Param(Array("10", "50", "100", "1000", "5000", "10000"))
  var size: Int = _

  var set: HashSet[Int] = _
  var list: List[Int] = _

  @Setup(Level.Trial)
  def init(): Unit = {
    set = (1 to size).to(HashSet)
    val from = size / 2
    list = (from to (from + size)).to(List)
  }

  @Benchmark
  def concatWithList(bh: Blackhole): Unit = {
    bh.consume(set.concat(list))
  }
}
```

```

// with content reuse

[info] # Run complete. Total time: 00:06:05
[info] Benchmark                                                   (size)  Mode  Cnt        Score       Error   Units
[info] HashSetBenchmark.concatWithList.test                                       10  avgt   20      296.023 ±    10.338   ns/op
[info] HashSetBenchmark.concatWithList.test:·gc.alloc.rate                        10  avgt   20     1634.008 ±    52.980  MB/sec
[info] HashSetBenchmark.concatWithList.test:·gc.alloc.rate.norm                   10  avgt   20      760.000 ±     0.001    B/op
[info] HashSetBenchmark.concatWithList.test:·gc.churn.PS_Eden_Space               10  avgt   20     1626.591 ±    92.472  MB/sec
[info] HashSetBenchmark.concatWithList.test:·gc.churn.PS_Eden_Space.norm          10  avgt   20      756.538 ±    35.456    B/op
[info] HashSetBenchmark.concatWithList.test:·gc.churn.PS_Survivor_Space           10  avgt   20        0.106 ±     0.032  MB/sec
[info] HashSetBenchmark.concatWithList.test:·gc.churn.PS_Survivor_Space.norm      10  avgt   20        0.050 ±     0.015    B/op
[info] HashSetBenchmark.concatWithList.test:·gc.count                             10  avgt   20      218.000              counts
[info] HashSetBenchmark.concatWithList.test:·gc.time                              10  avgt   20      179.000                  ms
[info] HashSetBenchmark.concatWithList.test                                       50  avgt   20     2033.927 ±    29.130   ns/op
[info] HashSetBenchmark.concatWithList.test:·gc.alloc.rate                        50  avgt   20     1095.090 ±    15.576  MB/sec
[info] HashSetBenchmark.concatWithList.test:·gc.alloc.rate.norm                   50  avgt   20     3504.003 ±     0.007    B/op
[info] HashSetBenchmark.concatWithList.test:·gc.churn.PS_Eden_Space               50  avgt   20     1101.024 ±    47.096  MB/sec
[info] HashSetBenchmark.concatWithList.test:·gc.churn.PS_Eden_Space.norm          50  avgt   20     3522.483 ±   132.999    B/op
[info] HashSetBenchmark.concatWithList.test:·gc.churn.PS_Survivor_Space           50  avgt   20        0.096 ±     0.026  MB/sec
[info] HashSetBenchmark.concatWithList.test:·gc.churn.PS_Survivor_Space.norm      50  avgt   20        0.306 ±     0.083    B/op
[info] HashSetBenchmark.concatWithList.test:·gc.count                             50  avgt   20      206.000              counts
[info] HashSetBenchmark.concatWithList.test:·gc.time                              50  avgt   20      156.000                  ms
[info] HashSetBenchmark.concatWithList.test                                      100  avgt   20     4134.369 ±    37.680   ns/op
[info] HashSetBenchmark.concatWithList.test:·gc.alloc.rate                       100  avgt   20     1039.193 ±     9.442  MB/sec
[info] HashSetBenchmark.concatWithList.test:·gc.alloc.rate.norm                  100  avgt   20     6760.007 ±     0.013    B/op
[info] HashSetBenchmark.concatWithList.test:·gc.churn.PS_Eden_Space              100  avgt   20     1034.846 ±    46.073  MB/sec
[info] HashSetBenchmark.concatWithList.test:·gc.churn.PS_Eden_Space.norm         100  avgt   20     6731.851 ±   295.540    B/op
[info] HashSetBenchmark.concatWithList.test:·gc.churn.PS_Survivor_Space          100  avgt   20        0.105 ±     0.041  MB/sec
[info] HashSetBenchmark.concatWithList.test:·gc.churn.PS_Survivor_Space.norm     100  avgt   20        0.684 ±     0.266    B/op
[info] HashSetBenchmark.concatWithList.test:·gc.count                            100  avgt   20      215.000              counts
[info] HashSetBenchmark.concatWithList.test:·gc.time                             100  avgt   20      172.000                  ms
[info] HashSetBenchmark.concatWithList.test                                     1000  avgt   20    60176.430 ±  1624.721   ns/op
[info] HashSetBenchmark.concatWithList.test:·gc.alloc.rate                      1000  avgt   20     1062.679 ±    27.278  MB/sec
[info] HashSetBenchmark.concatWithList.test:·gc.alloc.rate.norm                 1000  avgt   20   100536.100 ±     0.196    B/op
[info] HashSetBenchmark.concatWithList.test:·gc.churn.PS_Eden_Space             1000  avgt   20     1063.873 ±    65.900  MB/sec
[info] HashSetBenchmark.concatWithList.test:·gc.churn.PS_Eden_Space.norm        1000  avgt   20   100673.841 ±  5908.478    B/op
[info] HashSetBenchmark.concatWithList.test:·gc.churn.PS_Survivor_Space         1000  avgt   20        0.107 ±     0.036  MB/sec
[info] HashSetBenchmark.concatWithList.test:·gc.churn.PS_Survivor_Space.norm    1000  avgt   20       10.139 ±     3.330    B/op
[info] HashSetBenchmark.concatWithList.test:·gc.count                           1000  avgt   20      185.000              counts
[info] HashSetBenchmark.concatWithList.test:·gc.time                            1000  avgt   20      157.000                  ms
[info] HashSetBenchmark.concatWithList.test                                     5000  avgt   20   357200.388 ±  6501.661   ns/op
[info] HashSetBenchmark.concatWithList.test:·gc.alloc.rate                      5000  avgt   20      684.212 ±    12.338  MB/sec
[info] HashSetBenchmark.concatWithList.test:·gc.alloc.rate.norm                 5000  avgt   20   384392.590 ±     1.160    B/op
[info] HashSetBenchmark.concatWithList.test:·gc.churn.PS_Eden_Space             5000  avgt   20      676.084 ±    31.381  MB/sec
[info] HashSetBenchmark.concatWithList.test:·gc.churn.PS_Eden_Space.norm        5000  avgt   20   379963.682 ± 18364.617    B/op
[info] HashSetBenchmark.concatWithList.test:·gc.churn.PS_Survivor_Space         5000  avgt   20        0.249 ±     0.083  MB/sec
[info] HashSetBenchmark.concatWithList.test:·gc.churn.PS_Survivor_Space.norm    5000  avgt   20      139.996 ±    46.518    B/op
[info] HashSetBenchmark.concatWithList.test:·gc.count                           5000  avgt   20      170.000              counts
[info] HashSetBenchmark.concatWithList.test:·gc.time                            5000  avgt   20      149.000                  ms
[info] HashSetBenchmark.concatWithList.test                                    10000  avgt   20   847978.183 ±  8179.677   ns/op
[info] HashSetBenchmark.concatWithList.test:·gc.alloc.rate                     10000  avgt   20      782.462 ±    14.438  MB/sec
[info] HashSetBenchmark.concatWithList.test:·gc.alloc.rate.norm                10000  avgt   20  1043741.390 ± 14892.529    B/op
[info] HashSetBenchmark.concatWithList.test:·gc.churn.PS_Eden_Space            10000  avgt   20      793.203 ±    61.973  MB/sec
[info] HashSetBenchmark.concatWithList.test:·gc.churn.PS_Eden_Space.norm       10000  avgt   20  1058131.221 ± 82250.422    B/op
[info] HashSetBenchmark.concatWithList.test:·gc.churn.PS_Survivor_Space        10000  avgt   20        0.231 ±     0.117  MB/sec
[info] HashSetBenchmark.concatWithList.test:·gc.churn.PS_Survivor_Space.norm   10000  avgt   20      309.203 ±   158.327    B/op
[info] HashSetBenchmark.concatWithList.test:·gc.count                          10000  avgt   20      171.000              counts
[info] HashSetBenchmark.concatWithList.test:·gc.time                           10000  avgt   20      141.000                  ms





// without content reuse

[info] HashSetBenchmark.concatWithList.test                                       10  avgt   20      293.037 ±     4.908   ns/op
[info] HashSetBenchmark.concatWithList.test:·gc.alloc.rate                        10  avgt   20     1648.773 ±    26.838  MB/sec
[info] HashSetBenchmark.concatWithList.test:·gc.alloc.rate.norm                   10  avgt   20      760.000 ±     0.001    B/op
[info] HashSetBenchmark.concatWithList.test:·gc.churn.PS_Eden_Space               10  avgt   20     1646.644 ±    73.376  MB/sec
[info] HashSetBenchmark.concatWithList.test:·gc.churn.PS_Eden_Space.norm          10  avgt   20      758.782 ±    26.752    B/op
[info] HashSetBenchmark.concatWithList.test:·gc.churn.PS_Survivor_Space           10  avgt   20        0.107 ±     0.032  MB/sec
[info] HashSetBenchmark.concatWithList.test:·gc.churn.PS_Survivor_Space.norm      10  avgt   20        0.049 ±     0.014    B/op
[info] HashSetBenchmark.concatWithList.test:·gc.count                             10  avgt   20      237.000              counts
[info] HashSetBenchmark.concatWithList.test:·gc.time                              10  avgt   20      188.000                  ms
[info] HashSetBenchmark.concatWithList.test                                       50  avgt   20     2184.978 ±    29.835   ns/op
[info] HashSetBenchmark.concatWithList.test:·gc.alloc.rate                        50  avgt   20     1252.115 ±    17.060  MB/sec
[info] HashSetBenchmark.concatWithList.test:·gc.alloc.rate.norm                   50  avgt   20     4304.004 ±     0.007    B/op
[info] HashSetBenchmark.concatWithList.test:·gc.churn.PS_Eden_Space               50  avgt   20     1259.834 ±    60.236  MB/sec
[info] HashSetBenchmark.concatWithList.test:·gc.churn.PS_Eden_Space.norm          50  avgt   20     4330.034 ±   189.513    B/op
[info] HashSetBenchmark.concatWithList.test:·gc.churn.PS_Survivor_Space           50  avgt   20        0.104 ±     0.026  MB/sec
[info] HashSetBenchmark.concatWithList.test:·gc.churn.PS_Survivor_Space.norm      50  avgt   20        0.358 ±     0.091    B/op
[info] HashSetBenchmark.concatWithList.test:·gc.count                             50  avgt   20      192.000              counts
[info] HashSetBenchmark.concatWithList.test:·gc.time                              50  avgt   20      151.000                  ms
[info] HashSetBenchmark.concatWithList.test                                      100  avgt   20     4801.508 ±   979.006   ns/op
[info] HashSetBenchmark.concatWithList.test:·gc.alloc.rate                       100  avgt   20     1002.919 ±   144.133  MB/sec
[info] HashSetBenchmark.concatWithList.test:·gc.alloc.rate.norm                  100  avgt   20     7304.010 ±     0.022    B/op
[info] HashSetBenchmark.concatWithList.test:·gc.churn.PS_Eden_Space              100  avgt   20      998.299 ±   186.605  MB/sec
[info] HashSetBenchmark.concatWithList.test:·gc.churn.PS_Eden_Space.norm         100  avgt   20     7334.329 ±  1395.712    B/op
[info] HashSetBenchmark.concatWithList.test:·gc.churn.PS_Survivor_Space          100  avgt   20        0.070 ±     0.042  MB/sec
[info] HashSetBenchmark.concatWithList.test:·gc.churn.PS_Survivor_Space.norm     100  avgt   20        0.495 ±     0.275    B/op
[info] HashSetBenchmark.concatWithList.test:·gc.count                            100  avgt   20      134.000              counts
[info] HashSetBenchmark.concatWithList.test:·gc.time                             100  avgt   20      279.000                  ms
[info] HashSetBenchmark.concatWithList.test                                     1000  avgt   20    62491.229 ±  1521.718   ns/op
[info] HashSetBenchmark.concatWithList.test:·gc.alloc.rate                      1000  avgt   20     1208.714 ±    11.104  MB/sec
[info] HashSetBenchmark.concatWithList.test:·gc.alloc.rate.norm                 1000  avgt   20   118836.103 ±  2384.090    B/op
[info] HashSetBenchmark.concatWithList.test:·gc.churn.PS_Eden_Space             1000  avgt   20     1204.396 ±    33.801  MB/sec
[info] HashSetBenchmark.concatWithList.test:·gc.churn.PS_Eden_Space.norm        1000  avgt   20   118412.946 ±  3998.631    B/op
[info] HashSetBenchmark.concatWithList.test:·gc.churn.PS_Survivor_Space         1000  avgt   20        0.131 ±     0.035  MB/sec
[info] HashSetBenchmark.concatWithList.test:·gc.churn.PS_Survivor_Space.norm    1000  avgt   20       12.924 ±     3.581    B/op
[info] HashSetBenchmark.concatWithList.test:·gc.count                           1000  avgt   20      230.000              counts
[info] HashSetBenchmark.concatWithList.test:·gc.time                            1000  avgt   20      174.000                  ms
[info] HashSetBenchmark.concatWithList.test                                     5000  avgt   20   368192.274 ± 10214.897   ns/op
[info] HashSetBenchmark.concatWithList.test:·gc.alloc.rate                      5000  avgt   20      688.557 ±    18.378  MB/sec
[info] HashSetBenchmark.concatWithList.test:·gc.alloc.rate.norm                 5000  avgt   20   398528.627 ±     1.250    B/op
[info] HashSetBenchmark.concatWithList.test:·gc.churn.PS_Eden_Space             5000  avgt   20      691.393 ±    32.384  MB/sec
[info] HashSetBenchmark.concatWithList.test:·gc.churn.PS_Eden_Space.norm        5000  avgt   20   400199.555 ± 15772.738    B/op
[info] HashSetBenchmark.concatWithList.test:·gc.churn.PS_Survivor_Space         5000  avgt   20        0.269 ±     0.091  MB/sec
[info] HashSetBenchmark.concatWithList.test:·gc.churn.PS_Survivor_Space.norm    5000  avgt   20      155.459 ±    53.009    B/op
[info] HashSetBenchmark.concatWithList.test:·gc.count                           5000  avgt   20      200.000              counts
[info] HashSetBenchmark.concatWithList.test:·gc.time                            5000  avgt   20      167.000                  ms
[info] HashSetBenchmark.concatWithList.test                                    10000  avgt   20   842422.135 ± 25642.504   ns/op
[info] HashSetBenchmark.concatWithList.test:·gc.alloc.rate                     10000  avgt   20      788.200 ±    15.867  MB/sec
[info] HashSetBenchmark.concatWithList.test:·gc.alloc.rate.norm                10000  avgt   20  1043885.391 ± 14892.557    B/op
[info] HashSetBenchmark.concatWithList.test:·gc.churn.PS_Eden_Space            10000  avgt   20      803.332 ±    55.311  MB/sec
[info] HashSetBenchmark.concatWithList.test:·gc.churn.PS_Eden_Space.norm       10000  avgt   20  1063471.363 ± 65168.451    B/op
[info] HashSetBenchmark.concatWithList.test:·gc.churn.PS_Survivor_Space        10000  avgt   20        0.293 ±     0.131  MB/sec
[info] HashSetBenchmark.concatWithList.test:·gc.churn.PS_Survivor_Space.norm   10000  avgt   20      389.728 ±   174.027    B/op
[info] HashSetBenchmark.concatWithList.test:·gc.count                          10000  avgt   20      168.000              counts
[info] HashSetBenchmark.concatWithList.test:·gc.time                           10000  avgt   20      126.000                  ms
```